# docs/openapi.yml
openapi: 3.0.3
info:
  title: PR Reviewer Assignment Service (Test Task, Fall 2025)
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  parameters:
    TeamNameQuery:
      name: team_name
      in: query
      required: true
      schema:
        type: string
      description: Уникальное имя команды
    UserIdQuery:
      name: user_id
      in: query
      required: true
      schema:
        type: string
      description: Идентификатор пользователя
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - TEAM_EXISTS
                - PR_EXISTS
                - PR_MERGED
                - NOT_ASSIGNED
                - NO_CANDIDATE
                - NOT_FOUND
            message:
              type: string
      example:
        error:
          code: NOT_FOUND
          message: resource not found
...
    Team:
      type: object
      required: [ team_name, members]
      properties:
        team_name:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/TeamMember'
    User:
      type: object
      required: [ user_id, username, team_name, is_active ]
      properties:
        user_id:
          type: string
        username:
          type: string
        team_name:
          type: string
        is_active:
          type: boolean
    PullRequest:
      type: object
      required: [ pull_request_id, pull_request_name, author_id, status, assigned_reviewers]
      properties:
        pull_request_id:
          type: string
        pull_request_name:
          type: string
        author_id:
          type: string
        status:
          type: string
          enum: [OPEN, MERGED]
        assigned_reviewers:
          type: array
          items:
            type: string
          description: user_id назначенных ревьюверов (0..2)
        createdAt:
          type: string
          format: date-time
          nullable: true
        mergedAt:
          type: string
          format: date-time
          nullable: true
    PullRequestShort:
      type: object
      required: [ pull_request_id, pull_request_name, author_id, status]
      properties:
        pull_request_id:
          type: string
        pull_request_name:
          type: string
        author_id:
          type: string
        status:
          type: string
          enum: [OPEN, MERGED]
...
paths:
  /team/add:
    post:
      tags: [Teams]
      summary: Создать новую команду с участниками
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Team'
            example:
              team_name: payments
              members:
                - user_id: u1
                  username: Alice
                  is_active: true
                - user_id: u2
                  username: Bob
                  is_active: true
      responses:
        '201':
          description: Команда создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  team:
                    $ref: '#/components/schemas/Team'
              example:
                team:
                  team_name: backend
                  members:
                    - user_id: u1
                      username: Alice
                      is_active: true
                    - user_id: u2
                      username: Bob
                      is_active: true
        '400':
          description: Команда уже существует
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                error:
                  code: TEAM_EXISTS
                  message: team_name already exists
  /team/update:
    post:
      tags: [Teams]
      summary: Добавить или обновить участников существующей команды
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Team'
            example:
              team_name: backend
              members:
                - user_id: u3
                  username: Charlie
                  is_active: true
      responses:
        '200':
          description: Команда обновлена
          content:
            application/json:
              schema:
                type: object
                properties:
                  team:
                    $ref: '#/components/schemas/Team'
              example:
                team:
                  team_name: backend
                  members:
                    - user_id: u1
                      username: Alice
                      is_active: true
                    - user_id: u2
                      username: Bob
                      is_active: true
                    - user_id: u3
                      username: Charlie
                      is_active: true
        '404':
          description: Команда не найдена
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /team/get:
    get:
      tags: [Teams]
      summary: Получить команду с участниками
      parameters:
        - $ref: '#/components/parameters/TeamNameQuery'
      responses:
        '200':
          description: Объект команды
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
              example:
                team_name: backend
                members:
                  - user_id: u1
                    username: Alice
                    is_active: true
                  - user_id: u2
                    username: Bob
                    is_active: true
        '404':
          description: Команда не найдена
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /team/deactivateUsers:
    post:
      tags: [Teams]
      summary: Массовая деактивация пользователей команды с автоматическим переназначением открытых PR
      description: |
        Деактивирует указанных пользователей команды и автоматически переназначает
        все их открытые PR на других активных участников той же команды.
        Операция выполняется в одной транзакции для обеспечения консистентности.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ team_name, user_ids ]
              properties:
                team_name:
                  type: string
                  description: Имя команды
                user_ids:
                  type: array
                  items:
                    type: string
                  description: Список ID пользователей для деактивации
            example:
              team_name: backend
              user_ids: [u2, u3]
      responses:
        '200':
          description: Пользователи деактивированы и PR переназначены
          content:
            application/json:
              schema:
                type: object
                required: [ deactivated_users, reassigned_prs_count ]
                properties:
                  deactivated_users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                    description: Список деактивированных пользователей
                  reassigned_prs_count:
                    type: integer
                    description: Количество PR, которые были переназначены
              example:
                deactivated_users:
                  - user_id: u2
                    username: Bob
                    team_name: backend
                    is_active: false
                  - user_id: u3
                    username: Charlie
                    team_name: backend
                    is_active: false
                reassigned_prs_count: 5
        '404':
          description: Команда не найдена
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /users/setIsActive:
    post:
      tags: [Users]
      summary: Установить флаг активности пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ user_id, is_active ]
              properties:
                user_id:
                  type: string
                is_active:
                  type: boolean
            example:
              user_id: u2
              is_active: false
      responses:
        '200':
          description: Обновлённый пользователь
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
              example:
                user:
                  user_id: u2
                  username: Bob
                  team_name: backend
                  is_active: false
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /pullRequest/create:
    post:
      tags: [PullRequests]
      summary: Создать PR и автоматически назначить до 2 ревьюверов из команды автора
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ pull_request_id, pull_request_name, author_id ]
              properties:
                pull_request_id: { type: string }
                pull_request_name: { type: string }
                author_id: { type: string }
            example:
              pull_request_id: pr-1001
              pull_request_name: Add search
              author_id: u1
      responses:
        '201':
          description: PR создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  pr:
                    $ref: '#/components/schemas/PullRequest'
              example:
                pr:
                  pull_request_id: pr-1001
                  pull_request_name: Add search
                  author_id: u1
                  status: OPEN
                  assigned_reviewers: [u2, u3]
        '404':
          description: Автор/команда не найдены
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: PR уже существует
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                error: { code: PR_EXISTS, message: PR id already exists }
  /pullRequest/assignReviewers:
    post:
      tags: [PullRequests]
      summary: Автоматически назначить или дополнить ревьюверов для PR
      description: |
        Назначает ревьюверов автоматически из команды автора PR:
        - Если уже назначено 2 ревьювера - ничего не делает
        - Если назначен 1 ревьювер - добавляет второго
        - Если нет ревьюверов - назначает до 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ pull_request_id ]
              properties:
                pull_request_id: { type: string }
            example:
              pull_request_id: pr-1001
      responses:
        '200':
          description: Ревьюверы назначены или дополнены
          content:
            application/json:
              schema:
                type: object
                properties:
                  pr:
                    $ref: '#/components/schemas/PullRequest'
              example:
                pr:
                  pull_request_id: pr-1001
                  pull_request_name: Add search
                  author_id: u1
                  status: OPEN
                  assigned_reviewers: [u2, u3]
        '404':
          description: PR или автор не найден
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Нарушение доменных правил
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              examples:
                merged:
                  summary: Нельзя назначать ревьюверов для MERGED PR
                  value:
                    error: { code: PR_MERGED, message: cannot assign reviewers to merged PR }
  /pullRequest/merge:
    post:
      tags: [PullRequests]
      summary: Пометить PR как MERGED (идемпотентная операция)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ pull_request_id ]
              properties:
                pull_request_id: { type: string }
            example:
              pull_request_id: pr-1001
      responses:
        '200':
          description: PR в состоянии MERGED
          content:
            application/json:
              schema:
                type: object
                properties:
                  pr:
                    $ref: '#/components/schemas/PullRequest'
              example:
                pr:
                  pull_request_id: pr-1001
                  pull_request_name: Add search
                  author_id: u1
                  status: MERGED
                  assigned_reviewers: [u2, u3]
                  mergedAt: 2025-10-24T12:34:56Z
        '404':
          description: PR не найден
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /users/getReview:
    get:
      tags: [Users]
      summary: Получить PR'ы, где пользователь назначен ревьювером
      parameters:
        - $ref: '#/components/parameters/UserIdQuery'
      responses:
        '200':
          description: Список PR'ов пользователя
          content:
            application/json:
              schema:
                type: object
                required: [ user_id, pull_requests ]
                properties:
                  user_id:
                    type: string
                  pull_requests:
                    type: array
                    items:
                      $ref: '#/components/schemas/PullRequestShort'
              example:
                user_id: u2
                pull_requests:
                  - pull_request_id: pr-1001
                    pull_request_name: Add search
                    author_id: u1
                    status: OPEN
...