# internal/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// Defines values for ErrorResponseErrorCode.
const (
	NOCANDIDATE ErrorResponseErrorCode = "NO_CANDIDATE"
	NOTASSIGNED ErrorResponseErrorCode = "NOT_ASSIGNED"
	NOTFOUND    ErrorResponseErrorCode = "NOT_FOUND"
	PREXISTS    ErrorResponseErrorCode = "PR_EXISTS"
	PRMERGED    ErrorResponseErrorCode = "PR_MERGED"
	TEAMEXISTS  ErrorResponseErrorCode = "TEAM_EXISTS"
)

// Defines values for PullRequestStatus.
const (
	PullRequestStatusMERGED PullRequestStatus = "MERGED"
	PullRequestStatusOPEN   PullRequestStatus = "OPEN"
)

// Defines values for PullRequestShortStatus.
const (
	PullRequestShortStatusMERGED PullRequestShortStatus = "MERGED"
	PullRequestShortStatusOPEN   PullRequestShortStatus = "OPEN"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error struct {
		Code    ErrorResponseErrorCode `json:"code"`
		Message string                 `json:"message"`
	} `json:"error"`
}

// ErrorResponseErrorCode defines model for ErrorResponse.Error.Code.
type ErrorResponseErrorCode string

// PullRequest defines model for PullRequest.
type PullRequest struct {
	// AssignedReviewers user_id назначенных ревьюверов (0..2)
	AssignedReviewers []string          `json:"assigned_reviewers"`
	AuthorId          string            `json:"author_id"`
	CreatedAt         *time.Time        `json:"createdAt"`
	MergedAt          *time.Time        `json:"mergedAt"`
	PullRequestId     string            `json:"pull_request_id"`
	PullRequestName   string            `json:"pull_request_name"`
	Status            PullRequestStatus `json:"status"`
}

// PullRequestStatus defines model for PullRequest.Status.
type PullRequestStatus string

// PullRequestShort defines model for PullRequestShort.
type PullRequestShort struct {
	AuthorId        string                 `json:"author_id"`
	PullRequestId   string                 `json:"pull_request_id"`
	PullRequestName string                 `json:"pull_request_name"`
	Status          PullRequestShortStatus `json:"status"`
}

// PullRequestShortStatus defines model for PullRequestShort.Status.
type PullRequestShortStatus string

// ReviewerStatistics defines model for ReviewerStatistics.
type ReviewerStatistics struct {
	// AssignmentsCount Количество назначений на ревью
	AssignmentsCount int `json:"assignments_count"`

	// UserId Идентификатор пользователя
	UserId string `json:"user_id"`

	// Username Имя пользователя
	Username string `json:"username"`
}

// Team defines model for Team.
type Team struct {
	Members  []TeamMember `json:"members"`
	TeamName string       `json:"team_name"`
}

// TeamMember defines model for TeamMember.
type TeamMember struct {
	IsActive bool   `json:"is_active"`
	UserId   string `json:"user_id"`
	Username string `json:"username"`
}

// User defines model for User.
type User struct {
	IsActive bool   `json:"is_active"`
	TeamName string `json:"team_name"`
	UserId   string `json:"user_id"`
	Username string `json:"username"`
}

// TeamNameQuery defines model for TeamNameQuery.
type TeamNameQuery = string

// UserIdQuery defines model for UserIdQuery.
type UserIdQuery = string

// PostPullRequestAssignReviewersJSONBody defines parameters for PostPullRequestAssignReviewers.
type PostPullRequestAssignReviewersJSONBody struct {
	PullRequestId string `json:"pull_request_id"`
}

// PostPullRequestCreateJSONBody defines parameters for PostPullRequestCreate.
type PostPullRequestCreateJSONBody struct {
	AuthorId        string `json:"author_id"`
	PullRequestId   string `json:"pull_request_id"`
	PullRequestName string `json:"pull_request_name"`
}

// PostPullRequestMergeJSONBody defines parameters for PostPullRequestMerge.
type PostPullRequestMergeJSONBody struct {
	PullRequestId string `json:"pull_request_id"`
}

// PostPullRequestReassignJSONBody defines parameters for PostPullRequestReassign.
type PostPullRequestReassignJSONBody struct {
	OldReviewerId string `json:"old_reviewer_id"`
	PullRequestId string `json:"pull_request_id"`
}

// PostTeamDeactivateUsersJSONBody defines parameters for PostTeamDeactivateUsers.
type PostTeamDeactivateUsersJSONBody struct {
	// TeamName Имя команды
	TeamName string `json:"team_name"`

	// UserIds Список ID пользователей для деактивации
	UserIds []string `json:"user_ids"`
}

// GetTeamGetParams defines parameters for GetTeamGet.
type GetTeamGetParams struct {
	// TeamName Уникальное имя команды
	TeamName TeamNameQuery `form:"team_name" json:"team_name"`
}

// GetUsersGetReviewParams defines parameters for GetUsersGetReview.
type GetUsersGetReviewParams struct {
	// UserId Идентификатор пользователя
	UserId UserIdQuery `form:"user_id" json:"user_id"`
}

// PostUsersSetIsActiveJSONBody defines parameters for PostUsersSetIsActive.
type PostUsersSetIsActiveJSONBody struct {
	IsActive bool   `json:"is_active"`
	UserId   string `json:"user_id"`
}

// PostPullRequestAssignReviewersJSONRequestBody defines body for PostPullRequestAssignReviewers for application/json ContentType.
type PostPullRequestAssignReviewersJSONRequestBody PostPullRequestAssignReviewersJSONBody

// PostPullRequestCreateJSONRequestBody defines body for PostPullRequestCreate for application/json ContentType.
type PostPullRequestCreateJSONRequestBody PostPullRequestCreateJSONBody

// PostPullRequestMergeJSONRequestBody defines body for PostPullRequestMerge for application/json ContentType.
type PostPullRequestMergeJSONRequestBody PostPullRequestMergeJSONBody

// PostPullRequestReassignJSONRequestBody defines body for PostPullRequestReassign for application/json ContentType.
type PostPullRequestReassignJSONRequestBody PostPullRequestReassignJSONBody

// PostTeamAddJSONRequestBody defines body for PostTeamAdd for application/json ContentType.
type PostTeamAddJSONRequestBody = Team

// PostTeamDeactivateUsersJSONRequestBody defines body for PostTeamDeactivateUsers for application/json ContentType.
type PostTeamDeactivateUsersJSONRequestBody PostTeamDeactivateUsersJSONBody

// PostTeamUpdateJSONRequestBody defines body for PostTeamUpdate for application/json ContentType.
type PostTeamUpdateJSONRequestBody = Team

// PostUsersSetIsActiveJSONRequestBody defines body for PostUsersSetIsActive for application/json ContentType.
type PostUsersSetIsActiveJSONRequestBody PostUsersSetIsActiveJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Автоматически назначить или дополнить ревьюверов для PR
	// (POST /pullRequest/assignReviewers)
	PostPullRequestAssignReviewers(w http.ResponseWriter, r *http.Request)
	// Создать PR и автоматически назначить до 2 ревьюверов из команды автора
	// (POST /pullRequest/create)
	PostPullRequestCreate(w http.ResponseWriter, r *http.Request)
	// Пометить PR как MERGED (идемпотентная операция)
	// (POST /pullRequest/merge)
	PostPullRequestMerge(w http.ResponseWriter, r *http.Request)
	// Переназначить конкретного ревьювера на другого из его команды
	// (POST /pullRequest/reassign)
	PostPullRequestReassign(w http.ResponseWriter, r *http.Request)
	// Получить статистику назначений ревьюверов
	// (GET /statistics)
	GetStatistics(w http.ResponseWriter, r *http.Request)
	// Создать новую команду с участниками
	// (POST /team/add)
	PostTeamAdd(w http.ResponseWriter, r *http.Request)
	// Массовая деактивация пользователей команды с автоматическим переназначением открытых PR
	// (POST /team/deactivateUsers)
	PostTeamDeactivateUsers(w http.ResponseWriter, r *http.Request)
	// Получить команду с участниками
	// (GET /team/get)
	GetTeamGet(w http.ResponseWriter, r *http.Request, params GetTeamGetParams)
	// Добавить или обновить участников существующей команды
	// (POST /team/update)
	PostTeamUpdate(w http.ResponseWriter, r *http.Request)
	// Получить PR'ы, где пользователь назначен ревьювером
	// (GET /users/getReview)
	GetUsersGetReview(w http.ResponseWriter, r *http.Request, params GetUsersGetReviewParams)
	// Установить флаг активности пользователя
	// (POST /users/setIsActive)
	PostUsersSetIsActive(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Автоматически назначить или дополнить ревьюверов для PR
// (POST /pullRequest/assignReviewers)
func (_ Unimplemented) PostPullRequestAssignReviewers(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Создать PR и автоматически назначить до 2 ревьюверов из команды автора
// (POST /pullRequest/create)
func (_ Unimplemented) PostPullRequestCreate(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Пометить PR как MERGED (идемпотентная операция)
// (POST /pullRequest/merge)
func (_ Unimplemented) PostPullRequestMerge(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Переназначить конкретного ревьювера на другого из его команды
// (POST /pullRequest/reassign)
func (_ Unimplemented) PostPullRequestReassign(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Получить статистику назначений ревьюверов
// (GET /statistics)
func (_ Unimplemented) GetStatistics(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Создать новую команду с участниками
// (POST /team/add)
func (_ Unimplemented) PostTeamAdd(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Массовая деактивация пользователей команды с автоматическим переназначением открытых PR
// (POST /team/deactivateUsers)
func (_ Unimplemented) PostTeamDeactivateUsers(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Получить команду с участниками
// (GET /team/get)
func (_ Unimplemented) GetTeamGet(w http.ResponseWriter, r *http.Request, params GetTeamGetParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Добавить или обновить участников существующей команды
// (POST /team/update)
func (_ Unimplemented) PostTeamUpdate(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Получить PR'ы, где пользователь назначен ревьювером
// (GET /users/getReview)
func (_ Unimplemented) GetUsersGetReview(w http.ResponseWriter, r *http.Request, params GetUsersGetReviewParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Установить флаг активности пользователя
// (POST /users/setIsActive)
func (_ Unimplemented) PostUsersSetIsActive(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// PostPullRequestAssignReviewers operation middleware
func (siw *ServerInterfaceWrapper) PostPullRequestAssignReviewers(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostPullRequestAssignReviewers(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostPullRequestCreate operation middleware
func (siw *ServerInterfaceWrapper) PostPullRequestCreate(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostPullRequestCreate(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostPullRequestMerge operation middleware
func (siw *ServerInterfaceWrapper) PostPullRequestMerge(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostPullRequestMerge(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostPullRequestReassign operation middleware
func (siw *ServerInterfaceWrapper) PostPullRequestReassign(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostPullRequestReassign(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetStatistics operation middleware
func (siw *ServerInterfaceWrapper) GetStatistics(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetStatistics(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostTeamAdd operation middleware
func (siw *ServerInterfaceWrapper) PostTeamAdd(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostTeamAdd(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostTeamDeactivateUsers operation middleware
func (siw *ServerInterfaceWrapper) PostTeamDeactivateUsers(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostTeamDeactivateUsers(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTeamGet operation middleware
func (siw *ServerInterfaceWrapper) GetTeamGet(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTeamGetParams

	// ------------- Required query parameter "team_name" -------------

	if paramValue := r.URL.Query().Get("team_name"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "team_name"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "team_name", r.URL.Query(), &params.TeamName)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "team_name", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTeamGet(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostTeamUpdate operation middleware
func (siw *ServerInterfaceWrapper) PostTeamUpdate(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostTeamUpdate(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetUsersGetReview operation middleware
func (siw *ServerInterfaceWrapper) GetUsersGetReview(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetUsersGetReviewParams

	// ------------- Required query parameter "user_id" -------------

	if paramValue := r.URL.Query().Get("user_id"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "user_id"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "user_id", r.URL.Query(), &params.UserId)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "user_id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetUsersGetReview(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostUsersSetIsActive operation middleware
func (siw *ServerInterfaceWrapper) PostUsersSetIsActive(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostUsersSetIsActive(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/pullRequest/assignReviewers", wrapper.PostPullRequestAssignReviewers)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/pullRequest/create", wrapper.PostPullRequestCreate)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/pullRequest/merge", wrapper.PostPullRequestMerge)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/pullRequest/reassign", wrapper.PostPullRequestReassign)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/statistics", wrapper.GetStatistics)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/team/add", wrapper.PostTeamAdd)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/team/deactivateUsers", wrapper.PostTeamDeactivateUsers)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/team/get", wrapper.GetTeamGet)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/team/update", wrapper.PostTeamUpdate)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/users/getReview", wrapper.GetUsersGetReview)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/users/setIsActive", wrapper.PostUsersSetIsActive)
	})

	return r
}